+++ app.py
@@ -3,7 +3,7 @@ from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 import json
-from transformers import pipeline
+from src.model.ner import initialize_ner_model, process_text
 
 app = FastAPI()
 
@@ -11,17 +11,7 @@ app.mount("/static", StaticFiles(directory="static"), name="static")
 
 templates = Jinja2Templates(directory="templates")
 
-model_name = "ckiplab/bert-base-chinese-ner"
-ner_pipeline = pipeline("ner", model=model_name, grouped_entities=True)
-
-def convert_to_serializable(data):
-    if isinstance(data, list):
-        return [convert_to_serializable(item) for item in data]
-    elif isinstance(data, dict):
-        return {key: convert_to_serializable(value) for key, value in data.items()}
-    elif isinstance(data, (float, type(None))) or str(type(data)).find('float32') != -1:
-        return float(data) if data is not None else None
-    return data
+ner_pipeline = initialize_ner_model()
 
 @app.get("/", response_class=HTMLResponse)
 async def get_root(request: Request):
@@ -31,6 +21,11 @@ async def get_root(request: Request):
 async def process_ner(request: Request):
     data = await request.json()
     text = data.get("text", "")
-    results = ner_pipeline(text)
-    converted_results = convert_to_serializable(results)
+    converted_results = process_text(text, ner_pipeline)
+    return {"results": converted_results}
+
+@app.get("/test_ner")
+async def test_ner():
+    test_text = "我來自台灣，是一個民主國家。"
+    converted_results = process_text(test_text, ner_pipeline)
     return {"results": converted_results}
